api/requirements.txt
# OpenTelemetry 0.46 instrumentation still imports pkg_resources; setuptools 82+ removed it — pin below 82.
setuptools>=70.0.0,<82

fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
slowapi==0.1.9
nanoid==2.0.0
httpx==0.27.2
redis==5.1.1
aio-pika==9.4.3
opentelemetry-sdk==1.25.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-sqlalchemy==0.46b0
opentelemetry-instrumentation-redis==0.46b0
prometheus-fastapi-instrumentator==7.0.0
python-json-logger==2.0.7